review: wrap flagged bare returns, split flush batch tests (#274)
Round 3 P2 items. Every bare error propagation at the changed call
boundaries the review enumerated now carries its operation: the
insert/update/delete and batch write paths (validate-tables, main-row,
EAV, change-log steps), the statement builders' per-column-family
assignment collection, and the flush batch export step.
flush_batch_test.go (538 lines, pre-existing violation extended here)
is split along the touched seam: the mark-flushed and
before-export-hook suites move to flush_batch_mark_test.go (324 + 234).
Also drops the stray blank line at EOF in export_sql_mode_cases_test.go
flagged by git diff --check.
