+++ internal/cdc/export_sql_mode_cases_test.go
@@ -89,4 +89,3 @@ func exportSQLModeCases(rowID uuid.UUID) []exportSQLModeCase {
 		},
 	}
 }
-
